Make code build with stack
The quickCheck tests don't function well because
of a problem generating integers in the right
range. I used Arbitrary and a filter, but too many
examples are thrown out so the suite fails.

The lampWalk program fails to generate a solution.
Since it did produce a solution with the old compiler
and libraries, it ought to now, but I haven't debugged
it yet.